refactor: move read filter diffmap to home viewmodel
