+++ app/src/main/java/me/ash/reader/ui/page/home/flow/FlowPage.kt
@@ -128,7 +128,7 @@ fun FlowPage(
         LaunchedEffect(listState.isScrollInProgress) {
             if (!listState.isScrollInProgress) {
                 val firstItemIndex = listState.firstVisibleItemIndex
-                val diffMap = flowViewModel.diffMap
+                val diffMap = homeViewModel.diffMap
                 if (firstItemIndex < pagingItems.itemCount)
                     for (index in 0 until firstItemIndex) {
                         val item = pagingItems.peek(index)
@@ -168,7 +168,7 @@ fun FlowPage(
 
     DisposableEffect(pagingItems) {
         onDispose {
-            flowViewModel.commitDiff()
+            homeViewModel.commitDiff()
         }
     }
 
@@ -176,7 +176,7 @@ fun FlowPage(
         scope.launch {
             owner.lifecycle.eventFlow.collect {
                 if (it == Lifecycle.Event.ON_PAUSE) {
-                    flowViewModel.commitDiff()
+                    homeViewModel.commitDiff()
                 }
             }
         }
@@ -202,7 +202,7 @@ fun FlowPage(
             val id = article.article.id
             val isUnread = article.article.isUnread
 
-            with(flowViewModel.diffMap) {
+            with(homeViewModel.diffMap) {
                 if (contains(id)) remove(id)
                 else put(id, Diff(isUnread = !isUnread))
             }
@@ -418,7 +418,7 @@ fun FlowPage(
                     }
                     ArticleList(
                         pagingItems = pagingItems,
-                        diffMap = flowViewModel.diffMap,
+                        diffMap = homeViewModel.diffMap,
                         isShowFeedIcon = articleListFeedIcon.value,
                         isShowStickyHeader = articleListDateStickyHeader.value,
                         articleListTonalElevation = articleListTonalElevation.value,
